Improvements made on get_user_badges.php
The goal is to have the database of user badges only need to store the
badges that have been unlocked, and add in the other badges that have
been locked with javascript. This feels most of the way there with my
code. All the working code is still there, just commented out.