Created forgot password page
Forgot password page allows the user to enter an email to receive a
verification code. A verification code is sent to the email address if
the email address is associated with a registered username. Trying to
use jQuery to modify DOM on forgot_password page so that it changes from
an email form to where the user can submit their verification form.